# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Gerenciamento de Mensagens",
    "description": "API para gerenciar mensagens com autenticação JWT",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://virtserver.swaggerhub.com/JOAOPEDROCOSTAMACHAD/Biofy-API-Itens/1.0.0",
      "description": "SwaggerHub API Auto Mocking"
    },
    {
      "url": "http://localhost:2026",
      "description": "Servidor de desenvolvimento"
    }
  ],
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "Autenticação"
        ],
        "summary": "Autenticar usuário",
        "description": "Autentica um usuário e retorna um token JWT",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login bem sucedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/mensagem": {
      "get": {
        "tags": [
          "Mensagem"
        ],
        "summary": "Listar todas as Mensagens",
        "description": "Retorna uma lista de todos as mensagens cadastradas",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Token JWT para autenticação",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de mensagens",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Item"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Mensagem"
        ],
        "summary": "Criar nova Mensagem",
        "description": "Cria uma nova mensagem no sistema",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Token JWT para autenticação",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Item2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Item criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/mensagem/{id}": {
      "get": {
        "tags": [
          "Mensagem"
        ],
        "summary": "Buscar mensagem por ID",
        "description": "Retorna uma mensagem específica pelo ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da mensagem",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Token JWT para autenticação",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mensagem encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "404": {
            "description": "Mensagem não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Mensagem"
        ],
        "summary": "Atualizar Mensagem",
        "description": "Atualiza uma mensagem existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da mensagem",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Token JWT para autenticação",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Item2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Mensagem atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "404": {
            "description": "Mensagem não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Itens"
        ],
        "summary": "Deletar uma mensagem",
        "description": "Remove uma mensagem do sistema",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da mensagem",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Token JWT para autenticação",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Mensagem deletado com sucesso"
          },
          "404": {
            "description": "Mensagem não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "Login": {
        "required": [
          "email",
          "senha"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "123@teste.com"
          },
          "senha": {
            "type": "string",
            "example": "123"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        }
      },
      "Item": {
        "required": [
          "mensagem",
          "nome"
        ],
        "type": "object",
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "example": "c4261734-a3af-4e05-8f4b-6302411eec52"
          },
          "nome": {
            "type": "string",
            "example": "João"
          },
          "mensagem": {
            "type": "string",
            "example": "Olá Mundo!"
          },
          "created_at": {
            "type": "string",
            "example": "2024-12-06T23:56:13.093759Z"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "mensagem de erro"
          }
        }
      },
      "Item2": {
        "required": [
          "mensagem",
          "nome"
        ],
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "example": "Joana"
          },
          "mensagem": {
            "type": "string",
            "example": "Hello World!"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}